ble: updates so macOS works
