+++ platforms/ble/ble_client_adaptor.go
@@ -3,7 +3,6 @@ package ble
 import (
 	"fmt"
 	"log"
-	"strconv"
 	"sync"
 
 	"github.com/pkg/errors"
@@ -87,7 +86,7 @@ func (b *ClientAdaptor) Connect() (err error) {
 	b.addr.Set(b.Address())
 
 	// scan for the address
-	ch := make(chan bluetooth.ScanResult, 1)
+	ch := make(chan bluetooth.ScanResult)
 	err = b.adpt.Scan(func(adapter *bluetooth.Adapter, result bluetooth.ScanResult) {
 		if result.Address.String() == b.Address() {
 			b.adpt.StopScan()
@@ -223,26 +222,4 @@ func getBLEAdapter(impl string) (*bluetooth.Adapter, error) {
 	}
 
 	return currentAdapter, nil
-}
-
-func convertUUID(cUUID string) string {
-	switch len(cUUID) {
-	case 4:
-		// convert to full uuid from "22bb"
-		uid, e := strconv.ParseUint("0x"+cUUID, 0, 16)
-		if e != nil {
-			return ""
-		}
-
-		uuid := bluetooth.New16BitUUID(uint16(uid))
-		return uuid.String()
-
-	case 32:
-		// convert "22bb746f2bbd75542d6f726568705327"
-		// to "00001234-0000-1000-8000-00805f9b34fb"
-		return fmt.Sprintf("%s-%s-%s-%s-%s", cUUID[:8], cUUID[8:12], cUUID[12:16], cUUID[16:20],
-			cUUID[20:32])
-	}
-
-	return cUUID
 }
